Bug#20468712:FIELD_CHARSET->CSET->FILL IN FIELD_STRING::UNPACK CAUSE HEAP-BUFFER-OVERFLOW
Problem:
========
With binlog_row_image set to minimal and for certain
combination of columns as shown in bug page, unpacking a row
results in a crash.

Analysis:
========
In row based replication during the execution of an update
statement the before image is considered as read set and the
after image is considered as write set. These two sets are
packed as before image and after image. During this packing
the initial bytes of the packet are reserved for storing
bitmaps of columns that are being packed. Number of bytes
required to store the bitmap is calculated as shown below.

null_byte_count= (bitmap_bits_set(cols) + 7) / 8;

Before image's byte count is based on the read set columns
and after image's byte count is based on the write set.
After these reserved bytes actual field data is packed.
While unpacking the data, the received row_images are
pointed to the locations past these reserved bytes and
actual field data is extracted.

In the existing code while unpacking the data, for both the
before image and after image the read set columns are being
specified on slave. This causes the offsets to point to
incorrect byte locations and this results in a crash.

For example in the bug scenario while packing the data the
null_byte_count of before image is 1 and null_byte_count of
after image is 2. While unpacking the after image, since we
are using before image's values the row_data pointer is
incorrectly incremented by 1 pointer and the offset becomes
incorrect.

Fix:
===
Similar to packing procedure during unpacking, the after
image column bitmap is passed for unpacking the after image.