Bug#20709391 SCHED_GETCPU() IN DEFAULT GET_RND_INDEX() COULD NOT GENERATE RANDOM COUNTERS
The sched_getcpu() in default get_rnd_index() doesn't generate a
random number as we expected, so if some threads occupy the same
CPU, they would get the same counter. After the push of WL#8316
which makes default purge threads from 1 to 4, the 4 purge threads
would try to request the same slot in the sync_wait_array due to
the not random generator.

We replace existing get_rnd_index() with the RDTSC one, which
generates more random numbers, to reduce collision. With RDTSC,
we also make the for-loop in sync_array_get_and_reserve_cell()
work correctly, instead of trying the same slot over and over.
And we abandon sched_getcpu() totally, since it just generates
limited numbers.

RB: 8362
Reviewed-by: Marko <marko.makela@oracle.com>