Bug#20391552 CRASH DURING MEMCACHE WORKLOAD
PROBLEM

Cursor transaction (crsr_trx) is initially used for a read operation (get) ,
so the crsr_trx->read_only option is set as 1,and then the same transaction
is used for write (set) operation without resetting the crsr_trx->read_only = 0
,so it crashes in the innodb layer. Usually the background thread commits
the transaction and resets it,but in this case the connection got the
connection lock first and reused the transaction which leads to crash.

FIX

Check the status of crsr_trx->read_only before using the transaction for
write.If set commit the transaction before reusing.