Bug #20478242    MEMCAHCED SET COMMAND ACCEPTS NEGATIVE VALUES FOR EXPIRE TIME
PROBLEM

Inside memcached server the expire time is stored as unsigned int,so when the
user gives a negative value it is converted to a large number and is accepted.

FIX

Simple and easy fix is to restrict the maximum value of expire time to INT_MAX32
since any negative value will be greater than this and we can reject it.

Also removed useless assert in  innodb_flush_sync_conn() function