Bug#20316779 Invalid linestrings with one point are permitted Bug#20316794 Invalid polygons with rings of less than 4 points are permitted
Forbid linestrings that have only one point and polygons having a ring that has
less than 4 points unless it's a non-closed ring with 3 points(which will be
automatically closed internally), in all functions that creates a piece of
geometry data.

Those with multiple identical points are allowed as before and counted
as multiple points.