Bug#20109861: MYSQL 5.6 CRASHES WITH "MYSQLD GOT SIGNAL 11"               BY QUERY WITH MULTIPLE JOINS
ISSUE:
------
When a subquery is converted to semi-join, the query is
flattened and a single query is created. The outer-most
query's select_lex class does not keep track of the inner
query's condition and between count. This will create a
problem when there are large number of tables in the join
condition of the inner query (25 or more in this case).

Trunk & 5.7: Fixed by Roy as part of "WL#5275: Process
             subqueries in FROM clause in the same way as
             view"
5.5: Does not convert a subquery to semi-join. Hence problem
     does not occur.

SOLUTION:
---------
During the conversion from subquery to sem-join, make sure
that cond_count and between_count are added to the outer
query's select_lex.