Bug#20464737 SEMI_SYNC: SLAVE IO FAILS WITH ER1236 'SLAVE HAS MORE              GTIDS THAN THE MASTER HAS'
Slaves, especially semisync slaves, encountered ER1236 when
connecting to their masters. When connecting to the master
with COM_GTID_DUMP request, the slave sends all replicated
gtids to master and master checks if all the gtids(only
the gtids have master's uuid) are included in its gtid_executed.
However, there is a chance that a gtid is already in slave's
gtid_executed, but not in master's gtid_executed. Gtid is
added into gtid_executed after it is binlogged, so it is
possible that a transaction is replicated and applied on
slave, but not committed on master yet.

To fix it, master will check if all gtids are in the set of
(gtid_executed, owned_gtids).