Bug#20748660 THD->MDL_CONTEXT.OWNS_EQUAL_OR_STRONGER_LOCK | CREATE_INFO->TABLESPACE
A failing DISCARD/IMPORT TABLESPACE may leave the flag
THD::tablespace_op in an incorrect state, causing a
succeeding statement to fail. This patch introduces
a class utilizing RAII to make sure the flag is set
and reset correctly by allocating a local instance
of the class on the stack. The flag is reset by the
destructor when the instance goes out of scope.

Within the server, the THD::tablespace_op flag is used
only in 'lock_table_names()' to avoid locking the
tablespace name if this is a DISCARD or IMPORT tablespace
command (which only support file-per-table tablespaces).
This is the situation caught by this bug report.

The THD::tablespace_op flag is exported by means of the
'thd_tablespace_op()' function, which is part of the plugin
API. This function is used by the SEs Archive, Blackhole,
and InnoDB (partitioned and non-partitioned).

The usage in Archive and Blackhole is in the overriding
'handler::store_lock()' functions.

Within InnoDB, the 'thd_tablespace_op()' function is used
when opening a table, to suppress errors normally being
thrown when attempting to open a table with no tablespace.