Bug#20733540 Assertion failed: !table || (!table->read_set ...)
The problem exists for natural join/join using queries where one of
the tables is a derived table, but is accessed through a join nest.
In this case, mark_common_columns() checks for table reference being
is_view_or_derived(), or a field is present. But this test misses
the case above, since the derived table is underneath the join nest.

The fix is to have one check for a field, which is the simple case,
and use the Item pointer for all other cases.