Bug #20759613 	ALTER TABLE CAN CRASH THE SERVER
This is a regression caused by
  Commit: 7e7d64a7e34a5c141b4f9f534ddd8320f3461f9d [7e7d64a]
  Author: Mattias Jonsson <mattias.jonsson@oracle.com>
  Date: March 20, 2015 at 11:00:12 AM CDT
  Bug#20554858: INNODB: ADD TABLESPACE SUPPORT FOR NATIVE PARTITIONING
which was reviewed by me.  That patch split m_file_per_table into two
booleans, but did not use them correctly.

Here, those variables and related routines are renamed for clarity,
the correct variables are used, and m_use_file_per_table is fully
initialized in one place. The test added to create_tabespace.test
fails without this code fix in the same way that the bug reports
the failure in n_test.