Bug #20713559 ASSERT !HAS_DONE_RESERVATION, FSEG_ALLOC_FREE_PAGE_LOW()
Problem:

When a BLOB is stored in InnoDB, there is calculation to determine
the number of pages and then the number of extents.  Once the correct
number of extents are calculated, then that many extents are reserved.
But during the calculation of number of extents, we are not taking
into account the compressed pages.

Solution:

Calculate the number of extents, by taking into account whether the
pages are compressed or not.

rb#8402 approved by Sunny.