Bug#19585938 Crash in get_full_func_mm_tree with null item_field->table_ref
The problem may occur if we have a grouped query with a non-grouped
subquery that contains a reference to an aggregate function, e.g. in
the WHERE clause. The aggregate function must not reference any columns.

Example query:

SELECT (SELECT 1
        FROM t1
        WHERE SUM(1) < id
       )
FROM t1
GROUP BY col1+col2;

The range optimizer is attempting to optimize the predicate SUM(1) < id.
SUM(1) is represented by an Item_aggregate_ref object in get_mm_tree().
Since aggregation is performed in the outer query, this item is an outer
reference. get_mm_tree() will then call get_full_func_mm_tree() with
the real_item() of the Item_aggregate_ref object as predicand, which is
an Item_field. This Item_field represents a field in the temporary table
allocated for grouping the outer table. But since this table has no
assigned TABLE_LIST object, trying to calculate the map() for this table
fails.

However, this seems to be a legacy issue. Before WL#7540 was pushed,
table map was 1. But this is actually an outer reference, so it was
wrongly seen as a local table. Actually, being an outer reference, this
item is const during evaluation of the inner query, so this is not a
candidate for range optimizer analysis at all.

The fix is just to skip the call to get_full_func_mm_tree() if the
argument representing the Item_aggregate_ref is an outer reference.

The problem was also identified for other predicate types like BETWEEN.
Fix has been provided for this too, together with test cases.

(cherry picked from commit 6a71c3bf2c2f8f3e771173c38157fd42a6806cd3)