Bug#20748660 THD->MDL_CONTEXT.OWNS_EQUAL_OR_STRONGER_LOCK | CREATE_INFO->TABLESPACE
This is a followup refactoring after the first bugfix.

The usage of the 'THD::tablespace_op' flag has been reviewed, and
we see that the (indirect) usage in Archive and Blackhole is
inappropriate. At the time this code was written, the 'store_lock()'
function in InnoDB contained a code block that was evidently just
copied to Archive (and later, to Blackhole).

Thus, the remaining usage is in InnoDB (partitioned and non-
partitioned), and in the server itself. Thus, we do the following
refactoring:

1. Remove the 'THD::tablespace_op' flag.
2. Remove the call to 'thd_tablespace_op()' from
   Archive and Blackhole.
3. Two new constants are added to Alter_info:
   'ALTER_DISCARD_TABLESPACE' and
   'ALTER_IMPORT_TABLESPACE'.
4. In sql_yacc.yy, when encountering a DISCARD or
   IMPORT statement, the bit corresponding to the
   constant defined above is set in 'Alter_info::flags'.
5. The 'thd_tablespace_op()' function is changed to
   use the 'Alter_info::flags' to decide whether we
   are doing a DISCARD or IMPORT, instead of the
   'THD::tablespace_op' flag. Additionally, it must
   verify that the SQL command is ALTER TABLE, because
   the 'Alter_info' is reset only when a new ALTER
   statement starts.
6. The 'm_tablespace_op' member and the enum
   'enum_tablespace_op_type' are removed from the
   class 'Sql_cmd_discard_import_tablespace'. Instead,
   the 'Alter_info::flags' is used to find the
   required information.