Bug#20683959 LOAD DATA INFILE IGNORES A SPECIFIC ROW SILENTLY UNDER DB CHARSET IS UTF8
READ_INFO::read_field has logic for reallocating the internal buffer
if it is too small. This did not work correctly for multibyte charsets.

Fix: If we are out of space, then push the current character on the
input stack, and break out of while loop, to reach realloc code.

Also: Ensure that we report an error before returning an error status.