Bug #20372749 FAILING ASSERTION: !MBMAXLEN || LEN >= MBMINLEN * (FIXED_LEN / MBMAXLEN)
Problem:

During ALTER TABLE, when the spatial index is populated, the size of the data
tuple is being calculated in the function
rec_get_converted_size_comp_prefix_low() for compressed row.  In this function
there is debug code to check the length of the fields.  These checks was not
taking into account the spatial index and the data type DATA_SYS_CHILD in the
node pointers.

Solution:

Give special treatment to the spatial index and the data type DATA_SYS_CHILD
in the node pointers.

rb#8091 approved by Jimmy.