Bug#20531812: MEMORY CONSUMED QUICKLY WHILE EXECUTING LOOP IN PROCEDURE
Problem:
========
With a long running procedure the thd::main_mem_root grows
very rapidly.

Analysis:
========
With binlog enabled, during commit of each instruction
set_trans_pos gets called which stores the current binary
log file name in a variable named "m_trans_fixed_log_file".
Memory for the variable is allocated from thd::main_mem_root
and it is set to NULL as part of THD::cleanup_after_query()
at the end of the instruction. The thd::main_mem_root also
gets freed at the end of the instruction. But a procedure
which inserts thousands of records, is in progrees,  the
thd::main_mem_root will be getting cleared only at the end
of the procedure. But for each instruction within the
procedure, memory will be allocated from main_mem_root and
the m_trans_fixed_log_file will be set to NULL at the end
of each instruction. This causes the main_mem_root to grow
very rapidly as we keep on allocating 512 bytes from the
main_mem_root.

Fix:
===
At present in case of stored functions and triggers we will
not reset this variable to NULL. We allocate memory once and
the same memory is reused. Hence a similar fix has been
implemented as part of stored procedures as well. There is
flag named "sp_runtime_ctx". This flag is set to true when
the execution context refers to stored procedure, stored
functions, triggers and events. So now when this flag is set
we will not reset the "m_trans_fixed_log_file" to NULL. The
memory will be allocated once from the main_mem_root and it
will be reused. Hence no memory growth.