Bug#19811953 ENVELOPE() FUNCTION RETURNS INVALID POLYGON Bug#20196720 ENVELOPE(P) RETURNS NULL FOR A NON NULL POLYGON P
Issues:
1. Sometimes the MBR of a geometry can degrade to a point or
vertical/horizontal line segment. For example, given a point P, its MBR
is the point P itself; given a horizontal or
vertical linestring L, its MBR is a horizontal/vertical line segment.
In both cases if we return a polygon in function st_envelope(), the polygon
must be a degenerated one and many GIS functions don't accept it, preventing
queries with nested function calls like below to be executed.

SELECT ST_AREA(ST_ENVELOPE(ST_GEOMETRYFROMTEXT(geom))) from t1;

2. ST_Envelope() returns NULL given an empty geometry collection. It should
return an empty geometry(represented in the form of an empty geometry
collection) in this case because the result isn't 'unknown', we know
it's an empty geometry.

3. ST_Envelope() returns NULL given a geometry having some points but are
geometrically invalid. But other parts of GIS expects a valid MBR in
this case.

Fix:
If an MBR degrades to a point/linesegment, return the point or
line segment respectively as result of ST_ENVELOPE() instead of an
invalid polygon;

In ST_ENVELOPE return an empty geometry collection given an empty
geometry collection.

If the geometry input is geometrically invalid but has a valid WKB string,
ST_Envelope() still calcs its MBR and returns a valid MBR.

Finally, we don't accept a linestring with one point or a polygon with
a ring of less than 4 points anywhere in MySQL GIS, this is also true
for ST_Envelope() because such GIS data is rejected by geometry constructor
functions such as ST_GeomFromText/WKB/GeoJson before reaching ST_Envelope.