Bug#20313067 CHECK TABLE REPORTS WRONG COUNT FOR SPATIAL INDEX AND OTHER GIS PROBLEMS
Updating spatial index, which take the delete-mark/insert way, causes
lots of dup recs in rtree, and they're mapping to the same cluster rec,
this cause this bug.
In fact, for spatial index updating, since the different geometry data
could generate same MBR, so, if the new index entry is same as old
entry, which means the MBR is not changed, we don't need do anything in
this case.

Reviewed-by: Jimmy Yang<jimmy.yang@oracle.com>
RB: 8493