BUG#20593808 - CRASH WITH PARTITIONED TABLES + MULTITABLE DELETE
Problem :
---------
When PK/(part of PK) is used as index scan we store the REF along
with row data per partition and try to fill up the REF from the record.
When the scan is on part of the PK and sql-layer is not asking for all
the columns of PK to be fetched, it causes access violation while
trying to extract the complete key from the record.

Solution :
----------
We don't need to store PK as REF since it is already there in ROW.
Identify the case when PK(clustered)is not there and store the ROW ID
along with the data which can be used to generate handler::ref when
::position is called and for secondary sort.

Reviewed-by: Mattias Jonsson <mattias.jonsson@oracle.com>

RB: 8332