Fix Bug#20568464 MITIGATE THE STRICTNESS OF STRICT_* MODES 			OF INNODB_CHECKSUM_ALGORITHM
Problem:

If strict_X checksum algorithm is set, and InnoDB reads a page from disk that
contains a valid checksum Y (!= X). Then it will crash the server and will
print a message as if the checksum is not valid.

Solution:

Print a message to the server error log and accept the page as valid
if the page checksum matches with innodb, crc32 or none checksum algorithm.

Warning message would look like:

innodb_checksum_algorithm is set to "strict_X" but the page contains a valid
checksum "Y". Accepting the page as valid. Change innodb_checksum_algorithm
to "X" to silently accept such pages or rewrite all pages so that they contain
"X" checksum.