Bug #20590548 UNABLE TO LOGIN WITH USER WHOSE PWD CHANGED FROM 5.6.23 MYSQLADMIN IN 5.7.6
Problem:
The semantics of the SET PASSWORD query changed in 5.7 and it does not support 'hash'
of the password. mysqladmin 5.6 always calculates a hash from the password and sends
it to server. 5.7 server reads hash as a plain text and calculates a hash again on
a hash (double hashing).

Solution:
5.6 mysqladmin checks version of the server. If the server is 5.7 or greater,
ALTER USER USER() IDENTIFIED BY '<plain_text>' query is used to change a password.
The warning is printed as well informing the user that the mysqladmin should be updated.

Reviewed-by: Georgi Kodinov <georgi.kodinov@oracle.com>