Test the mysys's lockfree hash
It is impossible to implement the ut_hash_interface_t using the mysys's
lockfree hash because it requires the called to maintain a "pins" object
on his own, in addition to the hash table itself. Thus, messup the neat
implementation in the unit test.

This commit is for historical reference and is going to be reverted.

The mysys's lockfree is about 10 times slower than the implementation in
ut0lock_free_hash.h for the workload in the ut0lock_free_hash-t unit
test.

WL#7170 InnoDB buffer estimates for tables and indexes