Fixed bug#20566243: ERROR WHILE DOING CREATE TABLE T1 SELECT
Block CREATE TABLE ... SELECT statement to create table to inherit
the generated column infos. If we do support,
e.g create table t(a int, b int as(a+1));
create table t1 as select b from t;
We have to add a into t1. So, we choose to ignore to inherit the gcol_info.