BUG#19729278: 5.7.5: NOISYNESS AND NO LOG THROTTLING (SQL COORDINATOR MESSAGES)
After the changes introduced by WL 6661, some notes
that were meant to be printed when log_warnings>1
started appearing in the error log. The information
regarding the multi-threaded applier in particular
was too verbose. Moreover, these notes exhibit a
very technical and development oriented nature. Most
of them are likely to not even make sense to the user
at all.

This patch replaces the calls to sql_print_information
with DBUG_PRINTs. This will remove the notes from the
error log but keep the messages around for debugging
purposes during the development cycle.