Bug#20443863 USE OF WORST_SEEKS IN FIND_BEST_REF() CAN LEAD TO WRONG QUERY PLAN
This patch fixes a bug in how the optimizer uses the
JOIN_TAB::worst_seeks variable when limiting the cost of a key
lookup. The variable JOIN_TAB::worst_seeks is initialized to be a cost
estimate for the maximum cost for a key lookup but it is used as an
estimate for the maximum number of blocks to read when calculating the
cost of key lookup in Optimize_table_order::find_best_ref(). If we
change the cost constant for IO block reads, this bug can impact the
cost of doing ref access versus other access methods and may produce
worse query plans.

Example: For a query like this:

  SELECT i2 FROM t1 WHERE i1 = 1 AND i2 = 1;

if we have an index on i1, this query can be executed using ref
access. But if we double the values for all cost constants, due to the
error in use of JOIN_TAB::worst_seeks, the cost for ref access is more
than doubled and can cause this query to be executed as a table scan.

The fix for this problem is to change the code in
Optimize_table_order::find_best_ref() to use JOIN_TAB::worst_seeks as
a cost estimate instead of a block estimate.