BUG#20753463 HANDLE_FATAL_SIGNAL (SIG=11) IN __STRLEN_SSE2_PMINUB ON              CHANGE MASTER
A server cannot be configured as a slave when it was started
with server_id not configured (as the default is 0) or configured
as 0.

The server was properly working when changes were requested to the
replication default channel, but was failing if the requested
change was for other channel than default.

Analysis
========

The server behavior previous to multi source replication was:

(a) A slave initialized with "server_id = 0" will not initialize the
    slave and will not allow CHANGE MASTER statements even after
    someone changes @@GLOBAL.server_id to a value greater than 0
    after startup.

(b) A slave initialized with "server_id > 0" will initialize the slave
    and will allow CHANGE MASTER statements even after someone changes
    @@GLOBAL.server_id to 0 after startup.

The multi source replication server was failing by not verifying if the
slave was initialized properly (server started with server_id > 0) before
trying to create a new replication channel.

Fix
===

Once initialized, a slave server will add the default channel master_info
into the msr_map. Without having its slave initialized, a server will have
an empty msr_map (msr_map.get_num_instances() == 0).

The server code already has a 'is_slave_configured()' function that relies
on the above condition. The change_master_cmd function now verifies if the
slave was configured before trying to issue the CHANGE MASTER statement.