rb8590 Fixes to sporadically failing on PB2 rpl_xa_survive_crash_debug
The test fails due to not following a valid pattern of crash-simulation
with the following restart.
It's fixed to follow the pattern.
Extensive run (1600 times) has not shown any regression now.

As a side effect a non-determinism of XA ROLLBACK is worked around
to be fixed in a separate bug fixing as commented.