Post-commit fix or work-around for Bug#20872655 debug instrumentation.
There appears to be a problem with WL#6860 that causes
trx_rollback_for_mysql() to be called from innobase_close_connection()
even when the transaction is in XA PREPARE state.

At this point, the DEBUG_SYNC would trigger an assertion failure,
because the being-disconnected THD object is not valid for DEBUG_SYNC
any more.

Workaround: Only execute the DEBUG_SYNC if the current statement
is XA ROLLBACK.

Reviewed-by: Jimmy Yang <jimmy.yang@oracle.com>