Bug#20615023 SIGNAL 11 IN ITEM_FIELD::RESULT_TYPE DURING 1ST EXECUTION OF PREPARED STMT
There is a correct optimization which replaces
"x IN (SELECT y FROM DUAL)"
with "x=y". But it ignores WHERE, so it also replaces
"x IN (SELECT y FROM DUAL WHERE condition)" with "x=y".
This is wrong, "condition" might be FALSE or UNKNOWN, so the subquery
should then produce a NULL result, and the predicate should be UNKNOWN.
In other words: presence of WHERE should turn the optimization off.
As all in-to-exists subquery code, this optimization is from 2003-2004;
a similar problem was found and fixed in 2006 (bug 24670) but this code
block was not handled.