Bug #20883256 	INNODB: WARNINGS: NONNULL PARAMETER WILL EVALUATE TO 'TRUE' ON FIRST ENCOUNTER
This compiler warning occurs on a MacOS10.10.2.
Use (var != NULL) in each affected assert.
Split each assert into individual asserts when '&&' was used.
Each attribute((nonnull)) are removed from affected functions
A few functions are reformatted to the new style, not all.
The if (found) will stay deleted because all the callers really
do send in a valid pointer. ut_ad(found != NULL); is added.
dict_set_corrupted_index_cache_only() is always used when
table == index->table. So only index is needed as a parameter
and the function can be greatly simplified.